requirements.txt
# Deep Learning Frameworks
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.15.0
tensorflow-gpu>=2.15.0

# Model Optimization
torch-pruning>=0.2.0
onnx>=1.10.0
onnxruntime>=1.8.0
tensorrt>=8.6.1
nvidia-pyindex>=1.0.9

# Dataset and Data Processing
pycocotools>=2.0.2
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.19.5
pandas>=1.3.3

# Drone Simulation
gymnasium>=0.29.1
stable-baselines3>=2.1.0
airsim>=1.8.1

# Utilities
tqdm>=4.62.3
matplotlib>=3.4.3
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Development Tools
black>=23.0.0
pylint>=3.0.0
pytest>=7.0.0

# Additional dependencies
pyyaml>=5.4.1
psutil>=5.8.0
GPUtil>=1.4.0
scikit-learn>=0.24.2
tensorboard>=2.6.0 